Refactor ref server into its own file
